Move DTO (Data Transfer Object) logic from Contact class to ContactDTO on Data layer.
