+++ Interview/AppConfiguration.swift
@@ -1,0 +1,9 @@
+//
+//  AppConfiguration.swift
+//  Interview
+//
+//  Created by Lucas Ferreira Machado on 30/08/21.
+//  Copyright © 2021 PicPay. All rights reserved.
+//
+
+import Foundation
