+++ Interview/Data/Services/ListContactsService.swift
@@ -2,19 +2,8 @@ import Foundation
 
 private let apiURL = "https://run.mocky.io/v3/1d9c3bbe-eb63-4d09-980a-989ad740a9ac"
 
-/*
- Json Contract
-[
-  {
-    "id": 1,
-    "name": "Shakira",
-    "photoURL": "https://api.adorable.io/avatars/285/a1.png"
-  }
-]
-*/
-
 class ListContactService {
-    func fetchContacts(completion: @escaping ([Contact]?, Error?) -> Void) {
+    func fetchContacts(completion: @escaping ([ContactDTO]?, Error?) -> Void) {
         guard let api = URL(string: apiURL) else {
             return
         }
@@ -27,7 +16,7 @@ class ListContactService {
             
             do {
                 let decoder = JSONDecoder()
-                let decoded = try decoder.decode([Contact].self, from: jsonData)
+                let decoded = try decoder.decode([ContactDTO].self, from: jsonData)
                 
                 completion(decoded, nil)
             } catch let error {
